lxos-hub/services/api/requirements.txt
fastapi
uvicorn[standard]
psycopg[binary]
pydantic[email]
python-dotenv
redis
rq
requests
pyyaml
python-multipart
